@@ -26,25 +26,25 @@
   "author": "Gil Pedersen <gpdev@gpost.dk>",
   "license": "BSD-2-Clause",
   "dependencies": {
+    "@hapi/bounce": "^1.3.0",
     "aws-sdk": "^2.179.0",
-    "bounce": "^1.2.0",
     "commander": "^2.3.0",
-    "debug": "^2.0.0",
-    "hls-segment-reader": "^4.0.1",
-    "m3u8parse": "^1.0.0",
-    "measured": "^1.0.0",
+    "debug": "^4.1.1",
+    "hls-segment-reader": "^5.0.0",
+    "m3u8parse": "^2.0.1",
+    "measured-core": "^1.41.0",
     "mime-types": "^2.0.1",
     "mkdirp": "^0.5.0",
     "noptify": "0.0.3",
-    "pati": "^1.1.0",
-    "readable-stream": "^2.0.2",
+    "pati": "^1.1.2",
+    "readable-stream": "^3.3.0",
     "stream-each": "^1.1.2",
-    "udp-blast": "^1.0.0",
-    "uristream": "^2.0.0",
+    "udp-blast": "^2.0.0",
+    "uristream": "^3.0.0",
     "write-file-atomic": "^2.3.0"
   },
   "devDependencies": {},
   "engines": {
-    "node": ">=8.9.0"
+    "node": ">=10.4.0"
   }
 }
