@@ -24,7 +24,7 @@ internals.KeyFetcher = function (uri, cookie) {
 
 internals.KeyFetcher.prototype.fetch = function () {
 
-  let key = new Buffer(0);
+  let key = Buffer.alloc(0);
   let headers = {};
   if (this.cookie) {
     headers.Cookie = this.cookie;
@@ -70,16 +70,18 @@ internals.fetchKey = function (keyUri, options) {
 };
 
 
-internals.getIdentityKey = function (keyAttrs) {
+internals.getIdentityKey = function (method, keyAttrs) {
 
-  for (let idx = 0; idx < keyAttrs.length; idx++) {
-    let key = keyAttrs[idx];
-    let keyformat = key.quotedString('keyformat');
-    if (!keyformat || keyformat === 'identity') {
+  for (const key of keyAttrs) {
+    const keyformat = key.quotedString('keyformat');
+    const keymethod = key.enumeratedString('method');
+    if (!(keyformat || keyformat === 'identity') && 
+        (keymethod === method || keymethod === 'NONE')) {
+
       return {
-        method: key.enumeratedString('method'),
-        uri: key.quotedString('uri'),
-        iv: key.hexadecimalInteger('iv')
+        method: keymethod,
+        uri: key.uri ? key.quotedString('uri') : undefined,
+        iv: key.iv ? key.hexadecimalInteger('iv') : undefined
       };
     }
   }
@@ -88,19 +90,18 @@ internals.getIdentityKey = function (keyAttrs) {
 };
 
 
-exports.decrypt = async function (stream, keyAttrs, options, next) {
+exports.decrypt = async function (stream, keyAttrs, options) {
 
   if (!keyAttrs || !options) {
     return stream;
   }
 
-  let key = internals.getIdentityKey(keyAttrs);
+  const key = internals.getIdentityKey('AES-128', keyAttrs);
   if (!key || key.method === 'NONE') {
     return stream;
   }
 
-  if (key.method !== 'AES-128' || !key.uri || !key.iv) {
-
+  if (!key.uri || key.iv === undefined) {
     // TODO: hard error when key is not recognized?
     throw new Error('unknown encryption parameters');
   }
@@ -115,7 +116,16 @@ exports.decrypt = async function (stream, keyAttrs, options, next) {
 
   let decrypt;
   try {
-    decrypt = Crypto.createDecipheriv('aes-128-cbc', keyData, key.iv);
+    // Convert to Buffer
+
+    const iv = Buffer.alloc(16);
+    let work = key.iv;
+    for (let i = 15; i >= 0 && work !== 0; --i) {
+      iv.writeUInt8(Number(work & 0xffn), i);
+      work = work >> 8n;
+    }
+
+    decrypt = Crypto.createDecipheriv('aes-128-cbc', keyData, iv);
   } catch (ex) {
     throw new Error('crypto setup failed: ' + (ex.stack || ex));
   }
