@@ -3,7 +3,7 @@
 const util = require('util');
 const debug = require('debug')('hls:tssmooth');
 
-const Transform = require('readable-stream/transform');
+const { Transform } = require('readable-stream');
 
 const internals = {};
 
@@ -61,7 +61,7 @@ function TsSmooth(options) {
 
   this.packetSize = options.packetSize || 7 * 188; // size of output packets
 
-  this.buffer = new Buffer(0);
+  this.buffer = Buffer.alloc(0);
 
   this.pcr = -1;
   this.last = null;
