@@ -1,8 +1,8 @@
 'use strict';
 
 const Util = require('util');
-const Readable = require('readable-stream/readable');
-const Passthrough = require('readable-stream/passthrough');
+const { Readable } = require('readable-stream');
+const { PassThrough } = require('readable-stream');
 
 const StreamEach = require('stream-each');
 const Oncemore = require('oncemore');
@@ -63,7 +63,7 @@ function HlsReader(segmentReader, options) {
 
   this.isReading = false;
   this.isHooked = false;
-  this.buffer = new Passthrough({ highWaterMark: this.bufferSize });
+  this.buffer = new PassThrough({ highWaterMark: this.bufferSize });
 
   const processStreams = () => {
 
@@ -184,9 +184,9 @@ HlsReader.prototype.hook = function hook() {
   this.emit('ready');
 };
 
-HlsReader.prototype.decrypt = function(stream, keyAttrs, next) {
+HlsReader.prototype.decrypt = function(stream, keyAttrs) {
 
-  return SegmentDecrypt.decrypt(stream, keyAttrs, { base: this.reader.baseUrl, key: this.key, cookie: this.cookie }, next);
+  return SegmentDecrypt.decrypt(stream, keyAttrs, { base: this.reader.baseUrl, key: this.key, cookie: this.cookie });
 };
 
 
