@@ -3,7 +3,7 @@
 const Path = require('path');
 const Url = require('url');
 
-const Bounce = require('bounce');
+const Bounce = require('@hapi/bounce');
 const Mime = require('mime-types');
 const StreamEach = require('stream-each');
 const M3U8Parse = require('m3u8parse');
@@ -67,17 +67,20 @@ HlsStreamRecorder.prototype.updateIndex = function(update) {
       this.index.version = Math.max(2, this.index.version);    // v2 is required to support the remapped IV attribute
       if (this.index.version !== update.version)
         debug('changed index version to:', this.index.version);
+
       this.index.segments = [];
       this.index.first_seq_no = this.seq;
       this.index.type = 'EVENT';
       this.index.ended = false;
       this.index.discontinuity_sequence = 0; // not allowed in event playlists
+
       if (!isNaN(this.startOffset)) {
         let offset = this.startOffset;
         if (!update.ended) {
           if (offset < 0) offset = Math.min(offset, -3 * this.index.target_duration);
         }
-        this.index.start.decimalInteger('time-offset', offset);
+
+        this.index.start.signedDecimalFloatingPoint('time-offset', offset);
       }
     }
     else {
@@ -112,9 +115,9 @@ HlsStreamRecorder.prototype.updateIndex = function(update) {
           variant.uri = rec.localUrl;
         });
 
-        let allGroups = [];
-        for (let group in this.index.groups)
-          Array.prototype.push.apply(allGroups, this.index.groups[group]);
+        const allGroups = [];
+        for (const groupItems of this.index.groups.values())
+          Array.prototype.push.apply(allGroups, groupItems);
 
         allGroups.forEach((groupItem, index) => {
 
@@ -146,7 +149,7 @@ HlsStreamRecorder.prototype.updateIndex = function(update) {
       }
       else {
         this.index.variants = [];
-        this.index.groups = {};
+        this.index.groups = new Map();
         this.index.iframes = [];
       }
     }
